# postbox-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Postamat Platform API
  version: 1.0.0-draft.1
  summary: Parcel locker platform for Türkmenpoçta — client app, kiosk terminal, admin panel.
  description: |
    Single source of truth for the HTTP contract between three front-ends
    (Flutter client app, kiosk terminal, admin panel) and the core backend.

    Conventions — see `docs/API-01.md` for prose:

    * All JSON fields are `snake_case`.
    * All identifiers are UUID (v7 preferred).
    * All timestamps are UTC, ISO-8601 with `Z`. Clients render in `Asia/Ashgabat`.
    * All money is an integer in minor units plus a currency code. Never a float.
    * All phone numbers are E.164 (`+993…`).
    * Breaking changes ship as `/api/v2`, never as edits to `/api/v1`.

    Generated clients: `orval`/`openapi-typescript` for web, `openapi-generator` for Dart.
    Do not hand-edit generated code — edit this file.
  contact:
    name: Postamat platform team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.postamat.tm/api/v1
    description: Production (inside Türkmenistan)
  - url: https://stage.postamat.tm/api/v1
    description: Staging
  - url: http://localhost:4010/api/v1
    description: Prism mock from this file
...
security:
  - clientAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth (client)]
      operationId: refreshToken
      summary: Rotate the token pair
      description: Refresh tokens are single-use. Reuse revokes the whole family.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New pair issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: 'Codes: `REFRESH_TOKEN_INVALID`, `REFRESH_TOKEN_EXPIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Auth (client)]
      operationId: logout
      summary: Revoke the current session
      description: Also removes the push token bound to this session, if any.
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/auth/login:
    post:
      tags: [Auth (admin)]
      operationId: adminLogin
      summary: Sign in to the admin panel
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  type: string
                  maxLength: 128
                password:
                  type: string
                  format: password
                  maxLength: 256
            example:
              login: admin@1337
              password: correct-horse-battery-staple
      responses:
        '200':
          description: |
            Authenticated. `permissions` drives which actions the UI renders —
            but the server checks them again on every call.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TokenPair'
                  - type: object
                    required: [user]
                    properties:
                      user:
                        $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 'Codes: `ADMIN_CREDENTIALS_INVALID`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Codes: `ADMIN_ACCOUNT_BLOCKED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Codes: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/auth/refresh:
    post:
      tags: [Auth (admin)]
      operationId: adminRefreshToken
      summary: Rotate the admin token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New pair issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: 'Codes: `REFRESH_TOKEN_INVALID`, `REFRESH_TOKEN_EXPIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/auth/logout:
    post:
      tags: [Auth (admin)]
      operationId: adminLogout
      summary: Revoke the current admin session
      security:
        - adminAuth: []
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/auth/me:
    get:
      tags: [Auth (admin)]
      operationId: getAdminMe
      summary: Current admin user, role and permissions
      security:
        - adminAuth: []
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ────────────────────────────────────────────────────────────── Auth (kiosk)
...
  /kiosk/staff/cells:
    get:
      tags: [Kiosk (staff)]
      operationId: listKioskStaffCells
      summary: All cells of this postamat, for staff mode only
      description: |
        The cell wall exists here and nowhere else. A member of staff needs to see the
        whole cabinet to unload expired parcels and to diagnose locks; the public does not.
      security:
        - staffAuth: []
      responses:
        '200':
          description: Cells with live state.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CellWithState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/cells:
    get:
      tags: [Admin — cells]
      operationId: adminListCells
      summary: Cells across all postamats — the monitor
      description: |
        One table for both screen `333:4809` and the live monitor. At 200 postamats
        this is ~8400 rows, so the client virtualises and always passes a `limit`.
        Cursor pagination, because rows change under the reader.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: postamat_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: 'Comma-separated `CellStatus` values, e.g. `free,booked`.'
          schema:
            type: string
            example: free,booked
        - name: cell_type_id
          in: query
          description: Comma-separated list.
          schema:
            type: string
        - name: city_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: |
            Cells with their current holder, if any. `held_by` is `null` for free cells.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CellWithState'
                  pagination:
                    $ref: '#/components/schemas/CursorPagination'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin — cells]
      operationId: adminCreateCell
      summary: Create a single cell
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellWrite'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cell'
        '409':
          description: |
            Codes: `CELL_NUMBER_DUPLICATE`, `CELL_POSITION_OCCUPIED`,
            `HARDWARE_ADDRESS_DUPLICATE`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/cells/{cell_id}:
    get:
      tags: [Admin — cells]
      operationId: adminGetCell
      summary: One cell
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/CellId'
      responses:
        '200':
          description: Cell.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellWithState'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin — cells]
      operationId: adminUpdateCell
      summary: Edit a cell
      description: |
        Number, grid position and hardware address are three different things and are
        edited independently — the door label, where it sits in the cabinet, and which
        board output drives its lock.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/CellId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellWrite'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cell'
        '409':
          description: |
            Codes: `CELL_NUMBER_DUPLICATE`, `CELL_POSITION_OCCUPIED`,
            `HARDWARE_ADDRESS_DUPLICATE`, `CELL_HAS_ACTIVE_BOOKING`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/cells/{cell_id}/block:
    post:
      tags: [Admin — cells]
      operationId: adminBlockCell
      summary: Block a cell
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/CellId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 3
                  maxLength: 500
      responses:
        '200':
          description: Blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellWithState'
        '422':
          description: 'Codes: `REASON_REQUIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/cells/{cell_id}/unblock:
    post:
      tags: [Admin — cells]
      operationId: adminUnblockCell
      summary: Unblock a cell
      description: |
        A blocked cell may still have a parcel in it, so the panel must ask what
        happened to it before the cell can go back into circulation. `parcel_fate` is
        mandatory and lands in the audit log.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/CellId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parcel_fate]
              properties:
                parcel_fate:
                  type: string
                  enum: [collected, moved_to_counter, cell_was_empty, returned_to_sender]
                  description: The answer to «посылку забрали?».
                note:
                  type: [string, 'null']
                  maxLength: 500
            example:
              parcel_fate: moved_to_counter
              note: Клиент не пришёл, посылка на стойке выдачи
      responses:
        '200':
          description: Unblocked and returned to `free`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellWithState'
        '422':
          description: 'Codes: `PARCEL_FATE_REQUIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/roles:
    get:
      tags: [Admin — users and roles]
      operationId: adminListRoles
      summary: Roles and their permissions
      security:
        - adminAuth: []
      responses:
        '200':
          description: Roles.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─────────────────────────────────────────────────────── Admin — audit log
...
components:
  securitySchemes:
    clientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client access token, 15 minutes. Renew with `/auth/refresh`.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token, 15 minutes. Renew with `/admin/auth/refresh`.
    deviceAuth:
      type: http
      scheme: bearer
      description: |
        Long-lived device token obtained at provisioning. Kept in the device keystore.
    staffAuth:
      type: http
      scheme: bearer
      description: Staff-mode token on a terminal. Short-lived, dies with inactivity.
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Mandatory on every state-changing POST. A retry over a flaky connection must
        not book two cells or send two SMS.
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous response. Feeds and logs only.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Sort:
      name: sort
      in: query
      description: 'Field name, `-` prefix for descending. Example: `-created_at`.'
      schema:
        type: string
        example: -created_at
    Query:
      name: q
      in: query
      description: Free-text search. Scope depends on the endpoint.
      schema:
        type: string
        maxLength: 200
...
    CellId:
      name: cell_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: |
        Malformed request. Codes: `VALIDATION_ERROR`, `IDEMPOTENCY_KEY_MISSING`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request body failed validation.
              details:
                fields:
                  phone: must match E.164 (+993…)
              trace_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4eff
    Unauthorized:
      description: |
        No or expired credentials. Codes: `UNAUTHORIZED`, `TOKEN_EXPIRED`,
        `TOKEN_INVALID`. **401 means refresh the token — never confuse it with 403.**
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        Authenticated but not permitted. Codes: `FORBIDDEN`. **Refreshing will not help.**
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 'Codes: `NOT_FOUND`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: |
        Codes: `INTERNAL_ERROR`. `trace_id` is what to quote when reporting it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: UTC, ISO-8601 with `Z`. Rendered in `Asia/Ashgabat` by clients.
      example: '2026-08-14T09:41:00Z'
    Phone:
      type: string
      pattern: '^\+993[0-9]{8}$'
      description: E.164. Turkmenistan only for now.
      example: '+99365123456'
...
    HardwareAddress:
      type: object
      description: |
        Which board output drives this lock. Distinct from both the door number and
        the grid position — confusing the three is how a courier ends up opening the
        wrong cell.
      required: [board, output]
      properties:
        board:
          type: integer
          minimum: 0
        output:
          type: integer
          minimum: 0

    # ── errors ────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      description: The single error envelope. Every non-2xx response uses this shape.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: English, for developers and logs. Not shown to end users.
            details:
              type: [object, 'null']
              additionalProperties: true
              description: Structured extras — `attempts_left`, `fields`, `alternatives`.
            trace_id:
              type: string
              description: Correlates with server logs and the audit trail.
      example:
        error:
          code: CELL_ALREADY_TAKEN
          message: This cell was taken by another booking.
          details:
            cell_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e70
          trace_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4eaa

    # ── pagination ────────────────────────────────────────────────────────
...
    CursorPagination:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
        truncated_at:
          type: [string, 'null']
          format: date-time
          description: |
            Audit log only. Set when the requested range reaches past the one-year hot
            window — everything older lives in cold storage and comes out via CSV export.

    # ── auth ──────────────────────────────────────────────────────────────
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          description: Always `Bearer`.
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds.
          example: 900
    Permission:
      type: string
      description: |
        Granular capability. The client hides actions it lacks; the server enforces
        them regardless. Not an exhaustive enum on purpose — roles are a mechanism.
      examples:
        - postamats.read
        - postamats.write
        - cells.block
        - cells.remote_open
        - clients.block
        - audit.export
        - users.manage
    Role:
      type: object
      required: [id, code, name, permissions]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: superadmin
        name:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    AdminUser:
      type: object
      required: [id, login, full_name, role, permissions, is_active]
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        postamat_ids:
          type: array
          description: Empty means unrestricted.
          items:
            type: string
            format: uuid
        is_active:
          type: boolean
        last_login_at:
          type: [string, 'null']
          format: date-time

    # ── client ────────────────────────────────────────────────────────────
...
    CellStatus:
      type: string
      description: |
        Five states.

        * `free` — available.
        * `booked` — paid for, parcel not inside yet.
        * `occupied` — parcel inside, awaiting collection.
        * `blocked` — an operator's decision, and only that. Failed PIN attempts
          cannot produce it: a wrong code matches no cell, so there is nothing to
          block. What locks out is the terminal.
        * `maintenance` — planned work. Deliberately not the same thing as `blocked`.
      enum: [free, booked, occupied, blocked, maintenance]
    CellType:
      type: object
      required: [id, code, name, width_mm, height_mm, depth_mm, is_active]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: small
        name:
          type: string
          example: Маленький
        width_mm:
          type: integer
        height_mm:
          type: integer
        depth_mm:
          type: integer
        is_active:
          type: boolean
...
    Cell:
      type: object
      required: [id, postamat_id, number, cell_type_id]
      properties:
        id:
          type: string
          format: uuid
        postamat_id:
          type: string
          format: uuid
        number:
          type: string
          description: |
            The label on the door: a plain number, `1`–`42` in the pilot cabinet.
            Unique **within its postamat**, not across the network. Not the primary
            key and not the grid position.
          example: '14'
        row:
          type: [integer, 'null']
        col:
          type: [integer, 'null']
        row_span:
          type: [integer, 'null']
        col_span:
          type: [integer, 'null']
        hardware_address:
          oneOf:
            - $ref: '#/components/schemas/HardwareAddress'
            - type: 'null'
        cell_type_id:
          type: string
          format: uuid
    CellWrite:
      type: object
      properties:
        postamat_id:
          type: string
          format: uuid
        number:
          type: string
          maxLength: 16
        row:
          type: [integer, 'null']
          minimum: 1
        col:
          type: [integer, 'null']
          minimum: 1
        row_span:
          type: [integer, 'null']
          minimum: 1
        col_span:
          type: [integer, 'null']
          minimum: 1
        hardware_address:
          oneOf:
            - $ref: '#/components/schemas/HardwareAddress'
            - type: 'null'
        cell_type_id:
          type: string
          format: uuid
    CellWithState:
      allOf:
        - $ref: '#/components/schemas/Cell'
        - type: object
          required: [status]
          properties:
            status:
              $ref: '#/components/schemas/CellStatus'
            cell_type:
              $ref: '#/components/schemas/CellType'
            postamat_number:
              type: string
            postamat_address:
              type: string
            held_by:
              type: [object, 'null']
              description: Who is holding this cell right now. Null when `free`.
              properties:
                booking_id:
                  type: string
                  format: uuid
                client_id:
                  type: string
                  format: uuid
                client_phone:
                  $ref: '#/components/schemas/Phone'
                expires_at:
                  $ref: '#/components/schemas/Timestamp'
            status_reason:
              type: [string, 'null']
            status_changed_at:
              $ref: '#/components/schemas/Timestamp'
            door_open:
              type: [boolean, 'null']

    # ── bookings ──────────────────────────────────────────────────────
...